feat: RatingEvent 적용

+++ src/main/java/org/example/mopl/content/event/RatingEvent.java
@@ -8,18 +8,18 @@ public class RatingEvent {
 
   public record IncreaseRatingEvent(
       Long contentId,
-      int rating
+      double rating
   ) {
-    public static IncreaseRatingEvent of(Long contentId, int rating) {
+    public static IncreaseRatingEvent of(Long contentId, double rating) {
       return new IncreaseRatingEvent(contentId, rating);
     }
   }
 
   public record DecreaseRatingEvent(
       Long contentId,
-      int rating
+      double rating
   ) {
-    public static DecreaseRatingEvent of(Long contentId, int rating) {
+    public static DecreaseRatingEvent of(Long contentId, double rating) {
       return new DecreaseRatingEvent(contentId, rating);
     }
   }
